requirements.txt
Django==5.1.7
djangorestframework==3.15.2
pandas==2.2.2
openpyxl==3.1.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0